05312025
refactor base from po final test, may 26,2025

+++ src/main/java/com/rmj/guanzongroup/sidebarmenus/controller/PurchaseOrder_EntryController.java
@@ -550,7 +550,7 @@ public class PurchaseOrder_EntryController implements Initializable, ScreenInter
                     }
 
                     for (int lnCntr = 0; lnCntr <= detailCount - 1; lnCntr++) {
-                        int quantity = (int) poPurchasingController.PurchaseOrder().Detail(lnCntr).getValue("nQuantity");
+                        double quantity = (int) poPurchasingController.PurchaseOrder().Detail(lnCntr).getValue("nQuantity");
                         String stockID = (String) poPurchasingController.PurchaseOrder().Detail(lnCntr).getValue("sStockIDx");
 
                         // If any stock ID is empty OR quantity is 0, show an error and prevent saving
